requirements.txt
fastapi
uvicorn
openai
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
fpdf
faiss-cpu
sentence-transformers